// frontend/src/api/generated/model/incidentResponseAvailableActionsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * IncidentHub API
 * REST API for managing service ownership, dependencies, and incident lifecycle
 * OpenAPI spec version: 1.0.0
 */

export type IncidentResponseAvailableActionsItem =
  (typeof IncidentResponseAvailableActionsItem)[keyof typeof IncidentResponseAvailableActionsItem]

export const IncidentResponseAvailableActionsItem = {
  ASSIGN: 'ASSIGN',
  START: 'START',
  RESOLVE: 'RESOLVE',
  CLOSE: 'CLOSE',
  REOPEN: 'REOPEN',
  CANCEL: 'CANCEL',
} as const